fixed imports, in get port summary used get_po_service function directly, added models in HomePortSummary to Port.py
